Update favicon (closes #2152)
Update favicon to use a black and white variant of the colored logo.

An SVG favicon which support dark theme as been added too (https://blog.tomayac.com/2019/09/21/prefers-color-scheme-in-svg-favicons-for-dark-mode-icons/). Unfortunately this favicon is not picked up by chrome because the just-the-doc theme is declaring a favicon at the top of the header (and it cannot be removed without overriding the full header.html template).