Hide oldest unmaintained releases cycles by default (#50)
This implementation hides only a single run of rows, at the very end, if they are all unmaintained. The criteria picked here for hiding unmaintained releases is one which is simple enough to understand and gives the least surprise to users (the table only gets appended, no new rows show up in the middle). Release cycles are also only hidden if there is at least 6 release cycles in total, with at least 3 of of them that can be hidden. There is no need to hide anything if the table is small or if the number of hidden releases is too low.

This implementation works great for most of the current products, but may be useless for products where there are old maintained release cycles (such as Microsoft products or Ubuntu).

A button has been used instead of a link because to improve accessibility (https://www.a11y-101.com/design/button-vs-link).

See #50 if you want to know more about the reason why this solution was implemented.