[graalvm] Add GraalVM for OpenJDK 17/20 and update description
See https://www.graalvm.org/faq/ for information about the new naming and release cadence.

Despite the announcement that only the latest release is supported, both GraalVM for JDK 17 and JDK 20 were released. It is not clear whether LTS-based GraalVM versions will receive LTS support, so I did not say anything about it in the description and did not mark the JDK 17 version as LTS. But in any case the description will have to be updated in the future to clarify that point.