[new-product] Add Graylog (#3334)
The support/EOL policy for the commercial offerings were not detailed. We may have some hints about their current commercial support policy (https://github.com/endoflife-date/endoflife.date/issues/2482#issuecomment-1629446651), but this is not mentioned on https://graylog.org/pricing/ and it looks too vague. If one day that changes we can just add a note in the description and/or use the extended support column.

Closes #2487 

---------

Co-authored-by: Nemo <commits@captnemo.in>