Override css/activation.scss.liquid (#3448)
Override activation.scss.liquid to improve build time as suggested in https://github.com/just-the-docs/just-the-docs/pull/1244#issuecomment-1660246728.
As explained in this comment, this is possible because we currently have a flat navigation.

This override worth well the inconvenient, as time taken by a full build (bundle exec jekyll clean && bundle exec jekyll build) is greatly reduced:

- JtD 0.5.4: 87.896 seconds
- JtD 0.6.0: 56.159 seconds
- JtD 0.6.0 + override: 27.042 seconds