Add a validation rule to make sure releases are properly ordered (#3961)
According to the contributing guide (https://github.com/endoflife-date/endoflife.date/blob/master/CONTRIBUTING.md) releases must be ordered from the newest (on top of the list) to the lowest. This updates product-data-validator to enforce that rule.

The `outOfOrder` option allows to override this behavior, but it is undocumented because it's usually not recommended.

Close #3960.